[ADD] modified BluetoothDistanceDetector.java and MainActivity.java Don't mind the java file WiFi distance detector - it was made before I realized how to switch on the Bluetooth adapter in Raspberry.
The problem is trying to get a list of devices that is connected to Rasp's bluetooth module. Ideally it should see the nearby devices with bluetooth or at least one paired device (like we can see them in smartphone).

Finally, please don't mind the code's style (like constants should be in constant class). Firstly, I need the code to work, then it can be refined according to your requests. Unless the reason for a problem is due not refining the code as per your requests. =)